[Chore] #70 - QuizState파일 추가 및 코드 이동 & RootView에 건쏠지식 코드 추가
