+++ playkuround-iOS/Models/Games/Quiz.swift
@@ -37,42 +37,3 @@ func load<T: Decodable>(_ fileName: String) -> T {
         fatalError("Couldn't parse \(fileName) as \(T.self):\n\(error)")
     }
 }
-
-
-enum BlockState {
-    case normal
-    case unable
-    case correct
-    case incorrect
-    
-    var image: BlockImage {
-        switch self {
-        case .normal: return .quizBlock
-        case .unable: return .quizUnableBlock
-        case .correct: return .quizCorrectBlock
-        case .incorrect: return .quizIncorrectBlock
-        }
-    }
-    
-    var numberImage: NumberImage {
-        switch self {
-        case .normal: return .gray
-        case .unable: return .lightGray
-        case .correct: return .white
-        case .incorrect: return .white
-        }
-    }
-}
-
-enum BlockImage: String {
-    case quizBlock
-    case quizUnableBlock
-    case quizCorrectBlock
-    case quizIncorrectBlock
-}
-
-enum NumberImage: String {
-    case gray
-    case lightGray
-    case white
-}
